+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - (Improvement) (ML) CronJob status update
 - (Improvement) (ML) Job Sidecar Shutdown
 - (Feature) (ML) Handler for Extension StatefulSet and Service
+- (Feature) (ML) Pod & Container Config
 
 ## [1.2.35](https://github.com/arangodb/kube-arangodb/tree/1.2.35) (2023-11-06)
 - (Maintenance) Update go-driver to v1.6.0, update IsNotFound() checks